+++ lib/catan/interactors/buy_city.rb
@@ -43,10 +43,6 @@ class BuyCity < Interactor
 
   private
 
-  def raise_invalid_turn(turn)
-    raise IllegalOperation, "Invalid turn for this operation: #{turn}"
-  end
-
   def raise_action_not_taken
     raise IllegalOperation, 'Action has not been taken'
   end