+++ lib/catan/interactors/buy_settlement.rb
@@ -48,10 +48,6 @@ class BuySettlement < Interactor
 
   private
 
-  def raise_invalid_turn(turn)
-    raise IllegalOperation, "Invalid turn for this operation: #{turn}"
-  end
-
   def raise_action_not_taken
     raise IllegalOperation, 'Action has not been taken'
   end