+++ lib/catan/interactors/gain_resources.rb
@@ -47,10 +47,6 @@ class GainResources < Interactor
 
   private
 
-  def raise_invalid_turn(turn)
-    raise IllegalOperation, "Invalid turn for this operation: #{turn}"
-  end
-
   def raise_action_already_taken
     raise IllegalOperation, 'Action already taken'
   end