+++ lib/catan/interactors/interactor.rb
@@ -19,4 +19,8 @@ class Interactor
   def raise_uninitialized
     raise IllegalOperation, 'Game not initialized'
   end
+
+  def raise_invalid_turn(turn)
+    raise IllegalOperation, "Invalid turn for this operation: #{turn}"
+  end
 end