+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 - Fixed GameSerializer and GameDeserializer after adding cities to game state
+- Truly fixed ability to create roads from spots that don't border. Previously it was fixed
+  only on the entity level, while what really matters is interactor level. 
 
 ## [0.10.0]
 ### Added